refactoring main logic v 1.3

+++ src/main/java/com/example/MyFamilyMoney/Controllers/AccountController.java
@@ -1,21 +1,18 @@
 package com.example.MyFamilyMoney.Controllers;
 
-import com.example.MyFamilyMoney.models.Account;
-import com.example.MyFamilyMoney.models.AccountType;
-import com.example.MyFamilyMoney.models.Counteragent;
-import com.example.MyFamilyMoney.models.User;
-import com.example.MyFamilyMoney.repo.AccountRepository;
-import com.example.MyFamilyMoney.repo.CounteragentRepository;
-import com.example.MyFamilyMoney.repo.UserRepository;
+import com.example.MyFamilyMoney.models.*;
+import com.example.MyFamilyMoney.repo.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.annotation.CurrentSecurityContext;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import javax.validation.Valid;
 import java.util.ArrayList;
 import java.util.Optional;
 
@@ -25,9 +22,13 @@ public class AccountController {
     private AccountRepository accountRepository;
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private OperationsRepository operationsRepository;
+    @Autowired
+    private TransfersRepository transfersRepository;
 
     @GetMapping("/account")
-    public String accountMain(@CurrentSecurityContext(expression="authentication?.name") String username, Model model) {
+    public String accountMain(@CurrentSecurityContext(expression = "authentication?.name") String username, Model model) {
         User user = userRepository.findByUsername(username);
         Iterable<Account> accounts = accountRepository.findAllByUser(user);
         model.addAttribute("accounts", accounts);
@@ -35,18 +36,21 @@ public class AccountController {
     }
 
     @GetMapping("/account/add")
-    public String accountAdd(Model model) {
+    public String accountAdd(Model model, Account account) {
         return "account-add";
     }
 
     @PostMapping("/account/add")
-    public String accountAdd(@RequestParam String name, @RequestParam AccountType type, @RequestParam String startBalance,@CurrentSecurityContext(expression="authentication?.name") String username, Model model) {
+    public String accountAdd(@Valid Account account, BindingResult bindingResult, @CurrentSecurityContext(expression = "authentication?.name") String username) {
+        if (bindingResult.hasErrors()) {
+            return "account-add";
+        }
         User user = userRepository.findByUsername(username);
-        double startBalanceDouble = Double.parseDouble(startBalance)*100;
-        long startBalanceLong = (long) startBalanceDouble;
-        Account account = new Account(name, type, user, startBalanceLong, startBalanceLong);
+        account.setStartBalance(account.getStartBalance() * 100);
+        account.setEndBalance(account.getStartBalance());
+        account.setUser(user);
         accountRepository.save(account);
-        return "redirect:/account";
+        return "redirect:/";
     }
 
     @GetMapping("/account/{id}/edit")
@@ -67,14 +71,24 @@ public class AccountController {
         account.setName(name);
         account.setType(type);
         accountRepository.save(account);
-        return "redirect:/account";
+        return "redirect:/";
     }
 
     @PostMapping("/account/{id}/remove")
     public String accountDelete(@PathVariable(value = "id") long id, Model model) {
+        String answer;
         Account account = accountRepository.findById(id).orElseThrow();
-        accountRepository.delete(account);
-        return "redirect:/account";
+        Iterable<Operations> operations = operationsRepository.findAllByAccount(account);
+        Iterable<Transfer> transfersReceipt = transfersRepository.findAllByAccountReceipt(account);
+        Iterable<Transfer> transfersSpending = transfersRepository.findAllByAccountSpending(account);
+        if (operations.iterator().hasNext() | transfersReceipt.iterator().hasNext() | transfersSpending.iterator().hasNext()) {
+            answer = "redirect:/errorDeleteAccount";
+        }
+        else {
+            accountRepository.delete(account);
+            answer = "redirect:/";
+        }
+        return answer;
 
     }
 }
